atualizado para não receber quantidade menor ou igual a zero

+++ src/projetoFinal/Produtos.java
@@ -17,7 +17,7 @@ class Cliente {
 
 class CarrinhoCompras{
 	String fruta;
-	int quantidade;
+	double quantidade;
 	double preco;
 	double total;
 	int totalItens;
